+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+FEATURES:
+
+* Support for changing the default Vault address and Kubernetes mount path via CLI flag to the vault-csi-provider binary
+
 BUGS:
 
 * Added missing error handling when transforming SecretProviderClass config to a Vault request [[GH-97](https://github.com/hashicorp/vault-csi-provider/pull/97)]